+++ app/core/emotion_analyzer.py
