+++ app/core/text_processor.py
