+++ app/utils/helpers.py
@@ -1,0 +1,5 @@
+def format_time(seconds):
+    """Convert seconds to MM:SS format"""
+    minutes = int(seconds // 60)
+    seconds = int(seconds % 60)
+    return f"{minutes:02d}:{seconds:02d}"
